[FEAT] ckks test 파일 생성, ckks 파일 추가 및 코드 수정
