+++ tests/test_fhe_ckks.py
@@ -1,0 +1,5 @@
+from autofix.fhe_ckks import test_client
+
+if __name__ == "__main__":
+    test_client()
+
